Add some extra sanity checking.
Check that the deployment type is BLUE_GREEN just as an extra
precaution. We don't want to delete an Auto Scaling Group that is in use
and failed an in-place upgrade.

Theoretically this can't happen anyway as we check for a load balancer being
attached to the auto scaling group but it doesn't hurt to have some
extra sanity, and means we're also certain that we're only ever acting
on deployments of the BLUE_GREEN type.